biblio-app/requirements.txt


alembic
async-timeout
backports-abc
blinker
certifi
cffi
charset-normalizer
click
cryptography
dominate

flask

Flask-Bootstrap
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
flask-redis
Flask-Script
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
idna
itsdangerous
Jinja2
livereload
Mako
MarkupSafe
meld3
mysqlclient
packaging
PyMySQL
python-dotenv
python-editor
redis
requests
shellescape
simplejson
singledispatch
six
SQLAlchemy
supervisor
tornado
typing_extensions
urllib3
uWSGI
visitor
Werkzeug
WTForms

# OpenTelemetry Core and Instrumentation
prometheus-client
opentelemetry-distro
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-instrumentation-sqlalchemy
flask_prometheus_metrics[flask]
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests
fluent-logger
